+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Packaging changes
+- New dependency: [pynput](https://pypi.org/project/pynput/)
+- New macOS dependency: [accessibility](https://pypi.org/project/accessibility)
+
+### Added
+- Ability to switch output location on runtime
+- Ability to type output directly
+
+## [0.15] - 2018-06-07
+### Packaging changes
 - The macOS .dmg is now officially supported
 
 ### Changed